Validation of new source inputs. Updated terminology to mitigation sources in line with upcoming changes to database import to support a Mitigation Source table.
Adds the mitigation source table and the import for soruces

Adds text-search column generation to the mitigations and mitigation uses.

Adds header on success page for the mitigation based on the 'display name' in the database.

Refactors full_search to support multiple search queries.
Adds a mitigation_use card to the search result.
Adds stub functions for searching mitigation and mitigation_use.