Implements the mitigation source and mitigation pages.
Adds a field to store a regex to validate the ID of a specific mitigation source.
Renames generic functions to function names that are specific to techniques to allow function names specific to mitigations.
Adds initial mitigation pages and loading for mitigation sources and mitigations.
Adds page for Mitigation Source which lists all the mitigations of a given source.
Adds success html pages for mitigation and mitigation sources.